chore: move app structure
