+++ app/views.py
@@ -1,7 +1,7 @@
 from django.contrib.auth.models import Group, User
 from rest_framework import permissions, viewsets
 
-from tasking_back.tasking_back.serializers import GroupSerializer, UserSerializer
+from app.serializers import GroupSerializer, UserSerializer
 
 
 class UserViewSet(viewsets.ModelViewSet):
