--- tasking_back/tasking_back/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class TaskingBackConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'tasking_back'
