+++ tasking_back/urls.py
@@ -2,7 +2,7 @@ from django.urls import include, path
 from rest_framework import routers
 from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
 
-from tasking_back.tasking_back import views
+from app import views
 
 router = routers.DefaultRouter()
 router.register(r"users", views.UserViewSet)
